requirements.txt
matplotlib
numpy
scipy
bottle>=0.11.6
pybuilder>=0.10.36
minimock>=1.2.8
tinydb>=3.2.1
redislite>=3.2.311









